move disable_packet_mode to pakets_misc
